re-architec metrics as interface
