+++ app/graphql/types/publication_type.rb
@@ -2,6 +2,7 @@
 
 class PublicationType < BaseObject
   implements DoiItem
+  implements MetricInterface
 
   field :datasets, PublicationDatasetConnectionWithMetaType, null: false, description: "Referenced datasets", connection: true, max_page_size: 100 do
     argument :first, Int, required: false, default_value: 25
